backend/requirements.txt
# Core
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6

# ML Stack
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.36.0
Pillow==10.0.0
numpy>=1.26.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0

# Dev
pytest==7.4.0
pydantic-settings==2.1.0
